refactor transaction modules
