+++ package.json
@@ -43,7 +43,7 @@
     "eslint-config-vue": "latest",
     "eslint-friendly-formatter": "^4.0.1",
     "eslint-import-resolver-webpack": "^0.10.1",
-    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-node": "^7.0.1",
     "eslint-plugin-vue": "latest",
     "inject-loader": "^4.0.1",
