+++ src/modules/transactions/index.js
@@ -3,6 +3,7 @@ import { getDefaultState } from './defaultState';
 import actions from './actions';
 import getters from './getters';
 
+
 export default {
   state: getDefaultState(),
   mutations,
